+++ cmd/check.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 
 	log "github.com/sirupsen/logrus"
+	"github.com/spf13/viper"
 	"github.com/tommi2day/gomodules/pwlib"
 
 	"github.com/spf13/cobra"
@@ -42,6 +43,8 @@ func init() {
 	})
 	checkCmd.Flags().StringP("special_chars", "s", defaultSpecials, "define allowed special chars")
 	checkCmd.Flags().StringP("profile", "p", defaultProfile, "set profile string as numbers of 'length Upper Lower Digits Special FirstcharFlag(0/1)'")
+	checkCmd.Flags().StringP("profileset", "P", "", "set profile to existing named profile set")
+	// checkCmd.MarkFlagsMutuallyExclusive("profileset", "profile")
 	RootCmd.AddCommand(checkCmd)
 }
 
@@ -51,9 +54,33 @@ func checkPassword(cmd *cobra.Command, args []string) error {
 	var err error
 	log.Debugf("Args:%v", args)
 	password := args[0]
-	s, _ := cmd.Flags().GetString("special_chars")
-	pwlib.SetSpecialChars(s)
+	s, _ := cmd.Flags().GetString("profileset")
+	ch, _ := cmd.Flags().GetString("special_chars")
 	p, _ := cmd.Flags().GetString("profile")
+	if s != "" {
+		log.Debugf("got parameter profileset=%s", s)
+		ps := viper.GetStringMapString("password_profiles." + s)
+		if len(ps) > 0 {
+			if v, ok := ps["profile"]; ok {
+				p = v
+				log.Debugf("got parameter profile from parameterset %s", p)
+			} else {
+				log.Debugf("parameterset profile definition not found")
+				return fmt.Errorf("parameterset profile definition not found")
+			}
+			if v, ok := ps["special_chars"]; ok {
+				ch = v
+				log.Debugf("got parameter special_chars from parameterset: %s", ch)
+			} else {
+				log.Debugf("parameter special_chars from parameterset not set, use default: %s", ch)
+			}
+		} else {
+			log.Debugf("profilesset %s not found", s)
+			return fmt.Errorf("profileset %s not found", s)
+		}
+	}
+
+	pwlib.SetSpecialChars(ch)
 	profile, err = setPasswordProfile(p)
 	if err != nil {
 		return err