+++ cmd/genpass.go
@@ -4,6 +4,7 @@ package cmd
 import (
 	"fmt"
 
+	"github.com/spf13/viper"
 	"github.com/tommi2day/gomodules/pwlib"
 
 	log "github.com/sirupsen/logrus"
@@ -24,23 +25,6 @@ var newCmd = &cobra.Command{
 	SilenceUsage: true,
 }
 
-func genpass(cmd *cobra.Command, _ []string) error {
-	log.Debugf("generate password called")
-	s, _ := cmd.Flags().GetString("special_chars")
-	pwlib.SetSpecialChars(s)
-	p, _ := cmd.Flags().GetString("profile")
-	profile, err := setPasswordProfile(p)
-	if err != nil {
-		return err
-	}
-	password, err := pwlib.GenPassword(profile.Length, profile.Upper, profile.Lower, profile.Digits, profile.Special, profile.Firstchar)
-	if err == nil {
-		fmt.Println(password)
-		return nil
-	}
-	return err
-}
-
 func init() {
 	// hide unused flags
 	newCmd.SetHelpFunc(func(command *cobra.Command, strings []string) {
@@ -54,6 +38,48 @@ func init() {
 		command.Parent().HelpFunc()(command, strings)
 	})
 	newCmd.Flags().StringP("special_chars", "s", defaultSpecials, "define allowed special chars")
-	newCmd.Flags().StringP("profile", "p", defaultProfile, "set profile string as numbers of 'length Upper Lower Digits Special FirstcharFlag(0/1)'")
+	newCmd.Flags().StringP("profile", "p", defaultProfile, "set profile string as numbers of 'Length Upper Lower Digits Special FirstIsCharFlag(0/1)'")
+	newCmd.Flags().StringP("profileset", "P", "", "set profile to existing named profile set")
+	// newCmd.MarkFlagsMutuallyExclusive("profileset", "profile")
 	RootCmd.AddCommand(newCmd)
+}
+
+func genpass(cmd *cobra.Command, _ []string) error {
+	log.Debugf("generate password called")
+	s, _ := cmd.Flags().GetString("profileset")
+	ch, _ := cmd.Flags().GetString("special_chars")
+	p, _ := cmd.Flags().GetString("profile")
+	if s != "" {
+		log.Debugf("got parameter profileset=%s", s)
+		ps := viper.GetStringMapString("password_profiles." + s)
+		if len(ps) > 0 {
+			if v, ok := ps["profile"]; ok {
+				p = v
+				log.Debugf("got parameter profile from parameterset %s", p)
+			} else {
+				log.Debugf("parameterset profile definition not found")
+				return fmt.Errorf("parameterset profile definition not found")
+			}
+			if v, ok := ps["special_chars"]; ok {
+				ch = v
+				log.Debugf("got parameter special_chars from parameterset: %s", ch)
+			} else {
+				log.Debugf("parameter special_chars from parameterset not set, use default: %s", ch)
+			}
+		} else {
+			log.Debugf("profilesset %s not found", s)
+			return fmt.Errorf("profileset %s not found", s)
+		}
+	}
+	pwlib.SetSpecialChars(ch)
+	profile, err := setPasswordProfile(p)
+	if err != nil {
+		return err
+	}
+	password, err := pwlib.GenPassword(profile.Length, profile.Upper, profile.Lower, profile.Digits, profile.Special, profile.Firstchar)
+	if err == nil {
+		fmt.Println(password)
+		return nil
+	}
+	return err
 }